image_comparator/requirements.txt
Pillow>=10.0
imagehash>=4.3
dashscope>=1.14
httpx>=0.27
openpyxl>=3.1
streamlit>=1.35
pandas>=2.0
pytest>=8.0
pytest-asyncio>=0.23
